+++ src/at/tomtasche/mapsracer/data/Cluster.java
@@ -1,4 +1,4 @@
-package at.tomtasche.mapsracer;
+package at.tomtasche.mapsracer.data;
 
 public class Cluster {
 	private final int x;
