+++ src/at/tomtasche/mapsracer/math/VectorMagic.java
@@ -1,4 +1,4 @@
-package at.tomtasche.mapsracer;
+package at.tomtasche.mapsracer.math;
 
 import java.util.Iterator;
 import java.util.Set;
